Add option to disable requiring Kivy header
Added an option for the user to disable requiring a Kivy header. Defaults to true and then changing it to false will disable the error message. Figuring out how the LSP can read the configurations was confusing at first, but the example online (https://code.visualstudio.com/docs/extensions/example-language-server) helped a lot.

Also updated the server & client dependencies and the vscode engine.